// src/app/services/github-integration/models/github-repository.d.ts
export interface GitHubRepository {
  archive_url:string;
  archived:string;
  assignees_url:string;
  blobs_url:string;
  branches_url:string;
  clone_url:string;
  collaborators_url:string;
  comments_url:string;
  commits_url:string;
  compare_url:string;
  contents_url:string;
  contributors_url:string;
  created_at:string;
  default_branch:string;
  deployments_url:string;
  description:string;
  disabled:string;
  downloads_url:string;
  events_url:string;
  fork:string;
  forks:string;
  forks_count:string;
  forks_url:string;
  full_name:string;
  git_commits_url:string;
  git_refs_url:string;
  git_tags_url:string;
  git_url:string;
  has_downloads:string;
  has_issues:string;
  has_pages:string;
  has_projects:string;
  has_wiki:string;
  homepage:string;
  hooks_url:string;
  html_url:string;
  id:string;
  issue_comment_url:string;
  issue_events_url:string;
  issues_url:string;
  keys_url:string;
  labels_url:string;
  language:string;
  languages_url:string;
  license:string;
  merges_url:string;
  milestones_url:string;
  mirror_url:string;
  name:string;
  node_id:string;
  notifications_url:string;
  open_issues:string;
  open_issues_count:string;
  owner:string;
  private:string;
  pulls_url:string;
  pushed_at:string;
  releases_url:string;
  size:string;
  ssh_url:string;
  stargazers_count:string;
  stargazers_url:string;
  statuses_url:string;
  subscribers_url:string;
  subscription_url:string;
  svn_url:string;
  tags_url:string;
  teams_url:string;
  trees_url:string;
  updated_at:string;
  url:string;
  watchers:string;
  watchers_count:string;
}